[#16] .Refactor moved cosmos specific response handling out into separate file
